anime_video_mvp/requirements.txt
torch
diffusers==0.30.0
transformers==4.44.2
accelerate
opencv-python
gradio>=5.0.0
safetensors
einops
pillow
numpy<2.0.0
imageio[ffmpeg]
huggingface_hub
